OGM-122 cassandra backend
Exclude tests that cassandra can't support, which in some cases requires moving them to separate files to prevent annotated POJOs being processed if they would result in unsupported mappings.
