+++ core/src/test/java/org/hibernate/ogm/backendtck/associations/collection/manytomany/ManyToManyTest.java
@@ -24,35 +24,6 @@ import org.junit.Test;
 public class ManyToManyTest extends OgmTestCase {
 
 	@Test
-	public void testUnidirectionalManyToMany() {
-		Session session = openSession();
-		Transaction tx = session.beginTransaction();
-
-		Student john = new Student( "john", "John Doe" );
-		Student kate = new Student( "kate", "Kate Doe" );
-		Student mario = new Student( "mario", "Mario Rossi" );
-
-		ClassRoom math = new ClassRoom( 1L, "Math" );
-		math.getStudents().add( john );
-		math.getStudents().add( mario );
-		ClassRoom english = new ClassRoom( 2L, "English" );
-		english.getStudents().add( kate );
-		math.getStudents().add( mario );
-
-		persist( session, math, english, john, mario, kate );
-		tx.commit();
-
-		assertThat( getNumberOfEntities( sessions ) ).isEqualTo( 5 );
-		assertThat( getNumberOfAssociations( sessions ) ).isEqualTo( expectedAssociationNumber() );
-		session.clear();
-
-		delete( session, math, english, john, mario, kate );
-
-		session.close();
-		checkCleanCache();
-	}
-
-	@Test
 	public void testBidirectionalManyToMany() {
 		Session session = openSession();
 		Transaction tx = session.beginTransaction();
@@ -175,8 +146,6 @@ public class ManyToManyTest extends OgmTestCase {
 				BankAccount.class,
 				Car.class,
 				Tire.class,
-				Student.class,
-				ClassRoom.class
 		};
 	}
 }
