+++ core/src/test/java/org/hibernate/ogm/backendtck/associations/compositeid/ReferencedCompositeIdTest.java
@@ -10,7 +10,10 @@ import static org.fest.assertions.Assertions.assertThat;
 
 import org.hibernate.Session;
 import org.hibernate.Transaction;
+import org.hibernate.ogm.utils.GridDialectType;
 import org.hibernate.ogm.utils.OgmTestCase;
+import org.hibernate.ogm.utils.SkipByGridDialect;
+
 import org.junit.Test;
 
 /**
@@ -18,6 +21,10 @@ import org.junit.Test;
  *
  * @author Gunnar Morling
  */
+@SkipByGridDialect(
+		value = { GridDialectType.CASSANDRA },
+		comment = "Director.Tournament list - bag semantics unsupported (no primary key)"
+)
 public class ReferencedCompositeIdTest extends OgmTestCase {
 
 	@Test
