+++ core/src/test/java/org/hibernate/ogm/backendtck/failure/ErrorSpiTest.java
@@ -48,6 +48,10 @@ import com.google.common.util.concurrent.ThreadFactoryBuilder;
  *
  * @author Gunnar Morling
  */
+@SkipByGridDialect(
+		value = { GridDialectType.CASSANDRA },
+		comment = "Cassandra always upserts, doesn't read-lock before write, doesn't support uniq constraint even on primary key except by explicit/slow CAS use"
+)
 public class ErrorSpiTest extends OgmTestCase {
 
 	private static ExecutorService executor;
